build: sync to uni-db 2.0.0 + enable mimalloc on all bench bins
- object_store 0.11->0.13 (Cargo.toml); roaring 0.11.3->0.11.4 (lock,
  Lance 7 needs RoaringBitmap: Eq)
- blob_store/s3.rs: use object_store::ObjectStoreExt (0.13 moved
  get/put/head/delete to ext trait) + ObjectMeta::size as u64
- examples/compare_facts.rs: recursion_limit=256 (deeper async layout)
- ner/onnx.rs: fix String + &String build error (title_case)
- mimalloc global_allocator on the 7 remaining bench bins
- config/schema.json: regenerated snapshot (adds 2.0 label 'state' field)
- bugs/UNI_DB_WORKAROUNDS.md: catalog re-verified vs 2.0.0

cargo check --workspace --all-targets green; nextest 618 passed.

+++ crates/uniko-store/src/blob_store/s3.rs
@@ -12,7 +12,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use object_store::aws::AmazonS3Builder;
 use object_store::path::Path as ObjPath;
-use object_store::{ObjectStore, PutPayload};
+use object_store::{ObjectStore, ObjectStoreExt, PutPayload};
 
 use crate::blob_store::{BlobStore, PutOutcome, fs_relative_path};
 use crate::error::{Result, UnikoError};
@@ -117,7 +117,7 @@ impl BlobStore for S3BlobStore {
         let key = self.key(content_id)?;
         // Idempotency: skip upload if HEAD shows matching size.
         if let Ok(meta) = self.inner.head(&key).await
-            && meta.size == bytes.len()
+            && meta.size == bytes.len() as u64
         {
             return Ok(PutOutcome {
                 bytes_inline: None,
